Set up Dependabot for project
This commit sets up Dependabot for the project by modifying the `dependabot.yml` file. The configuration file is set to use the second version of the Dependabot configuration syntax, as indicated by the `version: 2` line. The `updates` section, which specifies which package ecosystems to update and where the package manifests are located, is currently left blank and will need to be filled in later. The `schedule` section is set to have Dependabot check for updates on a weekly basis.