Renamed `DependsUpon` to `DependsUponAttribute`
In this commit, the class `DependsUpon` has been renamed to `DependsUponAttribute`. This change has been reflected in the class definition and the constructor. Additionally, the methods `__ProcessPropertyAttributes` and `__ProcessMethodAttributes` in the `ViewModelBase` class have been updated to use the renamed `DependsUponAttribute` class when getting custom attributes of a property and a method respectively.