Added `SettingAttribute` class and new settings in `Settings.cs`
Added a new class `SettingAttribute` in `SettingAttribute.cs` which is a custom attribute for annotating properties with four parameters: `groupName`, `header`, `type`, and `isVisible`. The `System` namespace import in `Settings.cs` was replaced with `DBison.Core.Attributes` to use this new class.

Several new settings were added to the `Settings` class, each annotated with the `SettingAttribute` for additional metadata. These settings include `FontSize`, `UseDarkMode`, `Timeout`, `Limit`, `UseTranslucentWindow`, `TranslucentOpacity`, `UIScaling`, `UseSessionStore`, `MaxBackupCount`, `MaxLogSize`, `ShowExtendedDatabaseInformation`, and `OpenQueryOnServerAdded`.

Some of these settings also have a `Range` attribute applied to specify the valid range of values. For instance, `FontSize` has a range of 10 to 100, and `UIScaling` has a range of 0.1 to 2.0.